Make StringInterpolation interleaving string literals with interpolations
Fixed #6565

Currently `"foo#{"bar"}baz"` is parsed as `StringLiteral` with joining
all parts.
It is bad for formatting, so this changes it is parsed as `StringInterpolation`.
Above example's `expressions` property is now `["foo", "bar", "baz"]`.

In addition, this changes `StringInterpolation` ensure to interleave
string literals with interpolations.
In other words, it ensures `expressions` odd-index elements are
string literals and even-index elements are interpolations.
For example `"#{123}"` is parsed as `StringInterpolation` and its
`expressions` property is `["", 123, ""]`.
It is good for formatting.

Other changes in this are for keeping back-compatibility and formatter
improvement.